services/calendar_bot/requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# AI / LLM
openai==1.12.0

# Google Calendar
google-api-python-client==2.111.0
google-auth==2.26.2

# 配置管理
pydantic==2.5.3
pydantic-settings==2.1.0

# 数据库
sqlalchemy==2.0.25

# 工具库
pytz==2023.3
requests==2.31.0
tenacity==8.2.3
